Reorganized wallet code
This will probably need some heavy linting
