+++ app/plugins/Wallet/index.html
@@ -88,6 +88,8 @@
 		</div>
 		
 		<!-- JS -->
-		<script type='text/javascript' src='js/wallet.js'></script>
+		<script src='js/helpers.js'></script>
+		<script src='js/locking.js'></script>
+		<script src='js/buttons.js'></script>
 	</body>
 </html>
